Add d2l-organization-metadata
First stage of breaking apart the `d2l-course-image-tile`, this contains just the name/semester/code and notifications. Strictly speaking this isn't a fully-entity-driven component yet, as it still takes a few other inputs. However, an upcoming API change in the LMS will make this information available there, which will at least reduce the API of `d2l-course-image-tile` to just `href` and perhaps `widget-settings-href` or something.